+++ packages/flutter/lib/plass_ui.dart
@@ -35,6 +35,7 @@ export 'src/components/animate_lighting/pl_animate_lighting.dart';
 export 'src/components/animate_marquee/pl_animate_marquee.dart';
 export 'src/components/animate_reveal/pl_animate_reveal.dart';
 export 'src/components/animate_rotate/pl_animate_rotate.dart';
+export 'src/components/animate_scramble/pl_animate_scramble.dart';
 export 'src/components/animate_shake/pl_animate_shake.dart';
 export 'src/components/animate_slide/pl_animate_slide.dart';
 export 'src/components/animate_typing/pl_animate_typing.dart';